Extract design system primitives
