+++ Sources/QuillCodeApp/WorkspaceSwiftUIView.swift
@@ -2,24 +2,6 @@ import SwiftUI
 import QuillCodeCore
 import QuillCodeTools
 
-enum QuillCodeMetrics {
-    static let minimumHitTarget: CGFloat = 40
-    static let toolCardMinimumHeight: CGFloat = 74
-    static let compactToolCardMinimumHeight: CGFloat = 58
-    static let toolCardHeaderHeight: CGFloat = 44
-    static let toolCardRawDetailsMaxHeight: CGFloat = 240
-    static let toolCardRadius: CGFloat = 20
-    static let pressScale: CGFloat = 0.96
-}
-
-func quillCodeWithAnimation(_ animation: Animation, reduceMotion: Bool, _ updates: () -> Void) {
-    if reduceMotion {
-        updates()
-    } else {
-        withAnimation(animation, updates)
-    }
-}
-
 public struct QuillCodeWorkspaceView: View {
     public var surface: WorkspaceSurface
     @Binding public var draft: String
@@ -3587,72 +3569,6 @@ private struct QuillCodeWorktreeRemoveDraft: Equatable {
     }
 }
 
-struct QuillCodePressableButtonStyle: ButtonStyle {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .scaleEffect(!reduceMotion && configuration.isPressed ? QuillCodeMetrics.pressScale : 1)
-            .animation(reduceMotion ? nil : .easeOut(duration: 0.15), value: configuration.isPressed)
-    }
-}
-
-private extension View {
-    func quillCodeSurface(
-        fill: Color,
-        radius: CGFloat,
-        stroke: Color = Color.white.opacity(0.08),
-        shadow: Bool
-    ) -> some View {
-        modifier(QuillCodeSurfaceModifier(
-            fill: fill,
-            radius: radius,
-            stroke: stroke,
-            shadow: shadow
-        ))
-    }
-
-    func quillCodeImageOutline(radius: CGFloat) -> some View {
-        modifier(QuillCodeImageOutlineModifier(radius: radius))
-    }
-}
-
-private struct QuillCodeSurfaceModifier: ViewModifier {
-    var fill: Color
-    var radius: CGFloat
-    var stroke: Color
-    var shadow: Bool
-
-    func body(content: Content) -> some View {
-        content
-            .background(fill)
-            .overlay(
-                RoundedRectangle(cornerRadius: radius, style: .continuous)
-                    .stroke(stroke, lineWidth: 1)
-            )
-            .clipShape(RoundedRectangle(cornerRadius: radius, style: .continuous))
-            .shadow(
-                color: shadow ? Color.black.opacity(0.18) : .clear,
-                radius: shadow ? 18 : 0,
-                x: 0,
-                y: shadow ? 10 : 0
-            )
-    }
-}
-
-private struct QuillCodeImageOutlineModifier: ViewModifier {
-    var radius: CGFloat
-
-    func body(content: Content) -> some View {
-        content
-            .clipShape(RoundedRectangle(cornerRadius: radius, style: .continuous))
-            .overlay(
-                RoundedRectangle(cornerRadius: radius, style: .continuous)
-                    .stroke(Color.white.opacity(0.10), lineWidth: 1)
-            )
-    }
-}
-
 extension AgentMode {
     var title: String {
         switch self {
@@ -3664,19 +3580,4 @@ extension AgentMode {
             return "Auto"
         }
     }
-}
-
-enum QuillCodePalette {
-    static let background = Color(red: 0.03, green: 0.06, blue: 0.08)
-    static let sidebar = Color(red: 0.07, green: 0.10, blue: 0.12)
-    static let panel = Color(red: 0.10, green: 0.15, blue: 0.18)
-    static let selection = Color.white.opacity(0.08)
-    static let text = Color(red: 0.93, green: 0.97, blue: 0.98)
-    static let muted = Color(red: 0.62, green: 0.69, blue: 0.72)
-    static let blue = Color(red: 0.25, green: 0.72, blue: 0.91)
-    static let green = Color(red: 0.32, green: 0.82, blue: 0.45)
-    static let red = Color(red: 1.0, green: 0.36, blue: 0.32)
-    static let yellow = Color(red: 0.97, green: 0.72, blue: 0.31)
-    static let coral = Color(red: 0.82, green: 0.42, blue: 0.37)
-    static let purple = Color(red: 0.58, green: 0.50, blue: 0.96)
 }
